Refactor flash message visitor
